[compiler] Add wrapper functions to wrap change-detection storage and loading from the memo cache
Summary: We may wish to perform some additional computation on values when they enter or exit the memo cache in change detection mode (e.g. make a deep copy, restore the original value). This builds support for doing so.

In addition, it drops the "ForDebugging" part of the flag name and makes it compatible with "disableMemoization": if memoization is disabled, we implement that by not restoring the old version of the value unless we're in a source-level memo block.

ghstack-source-id: e1aeb03d83bba1400676a6b707dfec0e0cdde238
Pull Request resolved: https://github.com/facebook/react/pull/30181